Update for review comments

+++ tests/setup-gcstress.sh
@@ -1,4 +1,4 @@
-#!/bin/bash
+#!/usr/bin/env bash
 
 #
 # This script should be located in coreclr/tests.
@@ -6,15 +6,15 @@
 
 function print_usage {
     echo ''
-	echo 'Download coredistool for GC stress testing'
-	echo ''
-	echo 'Command line:'
-	echo ''
-	echo './setup-gcstress.sh --outputDir=<coredistools_lib_install_path>'
-	echo ''
-	echo 'Required arguments:'
-	echo '	--outputDir=<path>		   : Directory to install libcoredistools.so'
-	echo ''
+    echo 'Download coredistool for GC stress testing'
+    echo ''
+    echo 'Command line:'
+    echo ''
+    echo './setup-gcstress.sh --outputDir=<coredistools_lib_install_path>'
+    echo ''
+    echo 'Required arguments:'
+    echo '  --outputDir=<path>         : Directory to install libcoredistools.so'
+    echo ''
 }
 
 # Argument variables
@@ -24,28 +24,28 @@ libInstallDir=
 verbose=0
 for i in "$@"
 do
-	case $i in
-		-h|--help)
-			exit $EXIT_CODE_SUCCESS
-			;;
-		-v|--verbose)
-			verbose=1
-			;;
-		--outputDir=*)
-			libInstallDir=${i#*=}
-			;;
-		*)
-			echo "Unknown switch: $i"
-			print_usage
-			exit $EXIT_CODE_SUCCESS
-			;;
-	esac
+    case $i in
+        -h|--help)
+            exit $EXIT_CODE_SUCCESS
+            ;;
+        -v|--verbose)
+            verbose=1
+            ;;
+        --outputDir=*)
+            libInstallDir=${i#*=}
+            ;;
+        *)
+            echo "Unknown switch: $i"
+            print_usage
+            exit $EXIT_CODE_SUCCESS
+            ;;
+    esac
 done
 
 if [ -z "$libInstallDir" ]; then
-	echo "--libInstallDir is required."
-	print_usage
-	exit $EXIT_CODE_EXCEPTION
+    echo "--libInstallDir is required."
+    print_usage
+    exit $EXIT_CODE_EXCEPTION
 fi
 
 # This script must be located in coreclr/tests.
@@ -55,33 +55,54 @@ dotnetCmd=${dotnetToolsDir}/dotnetcli/bin/dotnet
 packageDir=${scriptDir}/../packages
 jsonFilePath=${scriptDir}/project.json
 
+# Check tool directory
+if [ ! -e $dotnetToolsDir ]; then
+    echo 'Directory containing dotnet commandline does not exist:' $dotnetToolsDir 
+    exit 1
+fi
+if [ ! -e $dotnetCmd ]; then
+    echo 'donet commandline does not exist:' $dotnetCmd
+    exit 1
+fi
+
 # make package directory
 if [ ! -e $packageDir ]; then
-  mkdir -p $packageDir
+    mkdir -p $packageDir
 fi
 
 # make output directory
 if [ ! -e $libInstallDir ]; then
-  mkdir -p $libInstallDir
+    mkdir -p $libInstallDir
 fi
 
 # Write dependency information to project.json
-echo {	\
-  \"dependencies\": { \
-  \"Microsoft.NETCore.CoreDisTools\": \"1.0.0-prerelease-00001\" \
-  }, \
-  \"frameworks\": { \"dnxcore50\": { } } \
-  } > $jsonFilePath
+echo {  \
+    \"dependencies\": { \
+    \"Microsoft.NETCore.CoreDisTools\": \"1.0.0-prerelease-00001\" \
+    }, \
+    \"frameworks\": { \"dnxcore50\": { } } \
+    } > $jsonFilePath
 
 # Find runtime id 
 rid=`$dotnetCmd --version | grep 'Runtime Id:' | sed 's/^ *Runtime Id: *//g'`
 
 # Download the package
-$dotnetCmd restore $jsonFilePath --source https://dotnet.myget.org/F/dotnet-core/ --packages $packageDir --runtime $rid
+echo Downloading CoreDisTools package
+bash -c -x "$dotnetCmd restore $jsonFilePath --source https://dotnet.myget.org/F/dotnet-core/ --packages $packageDir --runtime $rid"
+
+# Get library path
+libPath=`find $packageDir | grep libcoredistools.so`
+if [ ! -e $libPath ]; then
+    echo 'Failed to locate the downloaded library'
+    exit 1
+fi
 
 # Copy library to output directory
-libPath=`find $packageDir | grep libcoredistools.so`
+echo 'Copy library:' $libPath '-->' $libInstallDir/libcoredistools.so
 cp -f $libPath $libInstallDir
 
 # Delete temporary files
 rm -rf $jsonFilePath
+
+# Return success
+exit 0
